# perf(ext/url): use Ada URL parser