Ada is fast URL parser used in Node.js. https://github.com/ada-url/ada

```
# this patch
target/release/deno run -A cli/bench/url_parse.js
time 2258 ms rate 4428697
time 2240 ms rate 4464285

# main
deno run -A cli/bench/url_parse.js
time 2855 ms rate 3502626
time 2835 ms rate 3527336
```