@lucacasonato 2023-04-04
And the point is not that ada is better or worse at WPT or anything else like that. Ada seems cool, and is specifically not my point of contention. The problem is that if there are URLs where ada and rust-url disagree on how they parse a string into the URL representation, then two different parts of the system might disagree on, for example, what the host name is. If the permissions system says the hostname is `x.com` and checks this, and the `fetch` implementation thinks the hostname is `y-x.com` and connects there, we have a problem.