@lemire 2023-04-11
@lucacasonato 

>  The problem is that if there are URLs where ada and rust-url disagree on how they parse a string into the URL representation (...)

Are there such URLs?